futureready-tagging/requirements.txt
openai==2.26.0
dropbox==12.0.2
pandas==2.3.3
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
python-dotenv==1.2.1
openpyxl==3.1.5
streamlit==1.44.1
